% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_apply.R
\name{df_apply}
\alias{df_apply}
\title{"Transform" a data frame}
\usage{
df_apply(.data, .f, .condition = is.numeric, .else = identity, ...)
}
\arguments{
\item{.data}{– a data frame}

\item{.f}{– a function}

\item{.condition}{– a function – the default should TRUE for numeric columns and false for other columns}

\item{.else}{– a function to apply when .condition() evaluates to FALSE}

\item{...}{– additional arguments that will be passed to .f()}
}
\value{
a data frame with specific change made to it.
}
\description{
"Transform" a data frame
}
