% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FruitData.R
\docType{data}
\name{FruitData}
\alias{FruitData}
\title{Fruit data}
\format{
A data frame with 208 rows and 14 variables:
*{Country}{which country}
*{Code}{Country code}
*{Fruit}{Total fruit consumption}
*{Bananas}{Bananas consumption}
*{Dates}{Dates comsumption}
*{Citrus}{Citrus comsumption}
*{Oranges}{Oranges comsumption}
*{Apples}{Apples comsumption}
*{Lemons}{Lemonscomsumption}
*{Grapes}{Grapes comsumption}
*{Grapefruit}{Grapefruit comsumption}
*{Pineapples}{Pineapples comsumption}
*{Plantains}{comsumption}
*{Other}{other fruit comsumption}
}
\source{
\url{https://ourworldindata.org/grapher/fruit-consumption-by-fruit-type?country=~OWID_WRL}
}
\usage{
FruitData
}
\description{
This data set contains countries and their fruit consumptions
}
\examples{
FruitMax <- max(FruitData$Fruit)
}
\keyword{datasets}
