% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pivot_Data_Country.R
\docType{data}
\name{Pivot_Data_Country}
\alias{Pivot_Data_Country}
\title{Fruit Consumption by country (Pivot_Data_Country)}
\format{
A data frame with 50 rows and 5 variables:
*{Country}{which country}
*{Code}{Country code}
*{FruitType}{Fruit Type}
*{Consumption}{Fruit consumption}
*{Fruit}{total fruit consumption by that country}
}
\source{
\url{https://ourworldindata.org/grapher/fruit-consumption-per-capita}
}
\usage{
Pivot_Data_Country
}
\description{
This data set contains countries and their fruit consumptions
}
\examples{
 AvgComsumption <- mean(Pivot_Data_Country$Consumption)
}
\keyword{datasets}
