% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_apply.R
\name{df_apply}
\alias{df_apply}
\title{"Transform" a data frame}
\usage{
df_apply(.data, .f, .condition = is.numeric, .else = identity, ...)
}
\arguments{
\item{.data}{– a data frame}

\item{.f}{– a function}

\item{.condition}{– a function – the default should TRUE for numeric columns and false for other columns}

\item{.else}{– a function to apply when .condition() evaluates to FALSE}

\item{...}{– additional arguments that will be passed to .f()}
}
\value{
a data frame with specific change made to it.
}
\description{
"Transform" a data frame
}
\examples{
n <- 5L
TestData <- dplyr::tibble( double = c(5.002, 3.221, 4.008, 5.112, 4.779),
                integer   = c(0, 10, 20, 30, 40),
                cap = LETTERS[1L:n],
                low = letters[1L:n])
df_apply(test_data, round, toupper, digits = 1)
TestData |>  df_apply(toupper, purrr::is_character)
TestData |> df_apply(round, is.numeric, toupper, digits = 2),
}
